adding reservation and event web serivces

+++ src/main/java/tn/esprit/piproject/services/EvenementService.java
@@ -3,14 +3,18 @@ package tn.esprit.piproject.services;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import tn.esprit.piproject.entities.Evenement;
+import tn.esprit.piproject.entities.ReservationEvent;
 import tn.esprit.piproject.repositories.IEvenementRepository;
+import tn.esprit.piproject.repositories.IReservationEventRepository;
 
 import java.util.List;
+import java.util.Set;
 
 @Service
 @AllArgsConstructor
 public class EvenementService implements IEvenementSerivce {
     private IEvenementRepository iEvenementRepository;
+    private IReservationEventRepository iReservationEventRepository;
 
     @Override
     public List<Evenement> getAllEvents() {
@@ -35,8 +39,22 @@ public class EvenementService implements IEvenementSerivce {
 
     @Override
     public void deleteEvent(long id) {
-        iEvenementRepository.deleteById(id);
+        // Récupérer l'événement à supprimer
+        Evenement event = iEvenementRepository.findById(id).orElse(null);
+        if (event != null) {
+            // Récupérer l'ID de l'événement
+            Long evenementId = event.getId();
+
+            // Supprimer toutes les réservations associées à cet événement
+            iReservationEventRepository.deleteByEvenementId(evenementId);
+
+            // Supprimer l'événement lui-même
+            iEvenementRepository.deleteById(id);
+        } else {
+            System.out.println("L'événement avec l'ID " + id + " n'a pas été trouvé.");
+        }
     }
+
